Restored Windows version, may have broken the Linux version.
git-svn-id: https://tpb3dtg.svn.sourceforge.net/svnroot/tpb3dtg@60 7c74f4f4-39d4-4077-a54d-10deffb394f4

+++ People/Locations.h
@@ -372,7 +372,7 @@ class LocationLineageString	{
 	WorldLocations*wloc;
 	char sep[locseparatorStringSize];
 	char bufr[loclineageStringBufferSize];
-	inline int Len(){	return(strlen(bufr));	}
+	inline int Len(){	return (int)strlen(bufr);	}
 	//unary suffix operator !: length of string
 	inline int operator!(){	this->Len();	}
 	inline int MaxCatLen(){	return(loclineageStringBufferSize-Len());	}
